Check field file validity after validating
